Bugfixing

+++ inc/pressure.h
@@ -1,1 +1,1 @@
-long bmp180_calpressure(struct BMP180_CALIBRATION *,long,short);
+long bmp180_calpressure(struct BMP180_CALIBRATION *,uint32_t,int16_t);
